Homework #8 part 3
