+++ js/photo-load.js
@@ -1,61 +1,29 @@
-import {
-  isEscEvent
-} from './util.js';
-import {
-  imgUploadPreview
-} from './effects-photo.js';
-import {
-  destroyImgFilter
-} from './effects-photo.js';
+const FILE_TYPES = ['gif', 'jpg', 'jpeg', 'png'];
 
-const uploadOverlay = document.querySelector('.img-upload__overlay');
-const uploadInput = document.querySelector('.img-upload__input');
-const closePhoto = document.querySelector('.img-upload__cancel');
-const body = document.querySelector('body');
-
-const effectNone = document.querySelector('#effect-none');
-const textHashtags = document.querySelector('.text__hashtags');
-const textDescription = document.querySelector('.text__description');
-
-const onPhotoEscPress = (evt) => {
-  if (isEscEvent(evt)) {
-    evt.preventDefault();
-    uploadOverlay.classList.add('hidden');
-    body.classList.add('modal-open');
-    destroyImgFilter();
-  }
-};
+const fileChooser = document.querySelector('#upload-file');
+const preview = document.querySelector('.img-upload__preview img');
 
 const photoLoad = () => {
-  uploadInput.addEventListener('change', () => {
-    uploadOverlay.classList.remove('hidden');
-    body.classList.add('modal-open');
-    document.addEventListener('keydown', onPhotoEscPress);
-    effectNone.checked = true;
-    textHashtags.value = '';
-    textDescription.value = '';
-  });
+  fileChooser.addEventListener('change', () => {
+    const file = fileChooser.files[0];
+    const fileName = file.name.toLowerCase();
 
-  const closeButton = () => {
-    uploadOverlay.classList.add('hidden');
-    body.classList.remove('modal-open');
-    document.removeEventListener('keydown', onPhotoEscPress);
-    imgUploadPreview.removeAttribute('style');
-  }
+    const matches = FILE_TYPES.some((it) => {
+      return fileName.endsWith(it);
+    });
 
-  closePhoto.addEventListener('click', () => {
-    closeButton();
-    destroyImgFilter();
-  })
+    if (matches) {
+      const reader = new FileReader();
 
-  closePhoto.addEventListener('keydown', (evt) => {
-    if (isEscEvent(evt)) {
-      closeButton();
-      destroyImgFilter();
+      reader.addEventListener('load', () => {
+        preview.src = reader.result;
+      });
+
+      reader.readAsDataURL(file);
     }
   });
-};
+}
 
 export {
   photoLoad
-};
+}
